+++ run_tests.sh
@@ -948,6 +948,12 @@ ptest_program "examples/round76_test.om" 23
 ptest_program "examples/round79_test.om" 28
 ptest_program "examples/round80_threading_test.om" 8
 ptest_program "examples/round81_thread_keywords_test.om" 6
+ptest_program "examples/round82_nested_namespace_test.om" 12
+ptest_program "examples/round83_ns_fnref_test.om" 5
+ptest_program "examples/round84_ns_struct_test.om" 6
+ptest_program "examples/round85_ns_enum_test.om" 7
+ptest_program "examples/round86_ns_global_test.om" 7
+ptest_program "examples/round87_ns_type_qualtype_test.om" 8
 flush_ptests
 
 echo ""